+++ app/src/main/java/com/example/recipe_provider/dto/RecipeEntity.java
@@ -1,19 +1,19 @@
 package com.example.recipe_provider.dto;
 
 public class RecipeEntity {
-    private int id;
+    private long id;
     private String name;
     private int remain;
     private int require;
 
-    public RecipeEntity(int id, String name, int remain, int require) {
+    public RecipeEntity(long id, String name, int remain, int require) {
         this.id = id;
         this.name = name;
         this.remain = remain;
         this.require = require;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
     public String getName() {
